reorganize photo components folders
