+++ src/photos/components/single/Lightbox/BottomNavigation.tsx
