+++ src/photos/components/single/Lightbox/Buttons.tsx
