+++ src/photos/components/single/Lightbox/MainImage.tsx
